app/requirements.txt
# Enterprise Telegram Bot Requirements
# Updated for refactored architecture

# Core Dependencies
aiolimiter==1.2.1
alembic==1.13.1
anyio==4.9.0
APScheduler==3.11.0
asyncio==3.4.3
asyncpg==0.30.0
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pycparser==2.22
requests==2.32.4
sniffio==1.3.1
socksio==1.0.0
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0

# Configuration and Environment
python-dotenv==1.1.1

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10  # For Alembic migrations (sync operations)

# Telegram Bot Framework
python-telegram-bot==22.2
telethon==1.39.0

# SMS Service
kavenegar==1.1.2

# Development and Testing (optional)
pytest==8.2.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
black==24.4.2
flake8==7.0.0
mypy==1.10.0
